refactor: add providers to separate folder
