+++ src/providers/GuidCompletionItemProvider.ts
