+++ src/providers/index.ts
@@ -1,0 +1,3 @@
+export * from "./GuidCodeLensProvider";
+export * from "./GuidCompletionItemProvider";
+export * from "./GuidHoverProvider";
